Kasli: merge lock into spiflash and share with flash.rs and config.rs
