+++ artiq/firmware/libboard_misoc/lib.rs
@@ -27,7 +27,7 @@ pub mod uart;
 #[cfg(has_spiflash)]
 pub mod spiflash;
 pub mod config;
-// #[cfg(feature = "uart_console")]
+#[cfg(feature = "uart_console")]
 #[macro_use]
 pub mod uart_console;
 #[cfg(all(feature = "uart_console", feature = "log"))]
@@ -44,4 +44,5 @@ pub mod io_expander;
 pub mod net_settings;
 #[cfg(has_slave_fpga_cfg)]
 pub mod slave_fpga;
+#[cfg(any(soc_platform = "kasli", soc_platform = "sayma_amc"))]
 pub mod flash;
