+++ artiq/firmware/runtime/mgmt.rs
@@ -1,7 +1,7 @@
 use log::{self, LevelFilter};
 
 use io::{Write, ProtoWrite, Error as IoError};
-use board_misoc::{config, boot, flash}; 
+use board_misoc::{config, flash}; 
 use board_misoc;
 use logger_artiq::BufferLogger;
 use mgmt_proto::*;
@@ -144,7 +144,6 @@ fn worker(io: &Io, stream: &mut TcpStream) -> Result<(), Error<SchedError>> {
                     Ok(())
                 })?;
             }
-
             Request::DebugAllocator =>
                 unsafe { println!("{}", ::ALLOC) },
 
@@ -155,36 +154,26 @@ fn worker(io: &Io, stream: &mut TcpStream) -> Result<(), Error<SchedError>> {
                         info!("Writing {} success", key);
                         Reply::RebootImminent.write_to(stream)
                     }
-                    Err(board_misoc::flash::Error::WriteFail{ sector }) => {
+                    Err(board_misoc::spiflash::Error::WriteFail{ sector }) => {
                         error!("Flash writing failed on {} in sector {}", key, sector);
                         Reply::Error.write_to(stream)
                     }
-                    Err(board_misoc::flash::Error::CorruptedFirmware) => {
+                    Err(board_misoc::spiflash::Error::CorruptedFirmware) => {
+                        error!("Corrupted firmware is not flashed");
                         Reply::CorruptedFirmware.write_to(stream)
+                    }
+                    Err(board_misoc::spiflash::Error::AlreadyLocked) => {
+                        error!("Attempt at reentrant access");
+                        Reply::Error.write_to(stream)
                     }
                     Err(_) => {
                         error!("Flash writing failed");
                         Reply::Error.write_to(stream)
                     } 
                 }?;
-
             }
-
             Request::Reload => {
-                match flash::reload() {
-                    Ok(_)  => {
-                        info!("Reload failed"); // cannot actually get here
-                        Reply::RebootImminent.write_to(stream)
-                    }
-                    Err(_) => {
-                        error!("Reload failed");
-                        Reply::Error.write_to(stream)
-                    }
-                }?;
-            }
-
-            Request::Awake => {
-                Reply::Success.write_to(stream)?;
+                flash::reload();
             }
         };
     }
